rwa 1.1 updates (#242)
* MERGE

* Azure.Identity

* resourceToken

* update DeploymentSettings

* update text

* set diagram

* update to new deploymentSettings type

* fix azd hook

* add diagram

* remove unused

* update diagram

* update dev diagram

* add dev diagram

* update diagram

* remove unused diagrams

* diagram - bring storage icon front

* fix private dns zone name length

* invalid deployment name trailing char

* update vnet diagram

* label as jumpbox

* fix dupe deployment names

* exceeds deployment name

* rearrange deployment name

* remove unused image

* Update file execution

* existing resource cannot use empty string

* upgrade

* sharedAzureContainerRegistry

* address provision error for ACR private endpoints

* update comment for clarity

* $hubGroup

* cleanup aca groups

* update vm sku

* add docker dependencies

* update description

* revert README

* use federated creds

* link to 3rd party doc

* update param descriptions

* Remove unnecessary code in cleanup.ps1

* new ACR module

* fix script validation

* use core module for dev

* update troubleshooting step

+++ testscripts/cleanup.ps1
@@ -21,6 +21,10 @@
     If you provide the ResourceGroup parameter and have deployed a hub network, then you must also provide
     the HubResourceGroup if it is a different resource group.  If you don't, then the hub network will not
     be cleaned up.
+.PARAMETER DeleteGroups
+    Defaults to true, but if you set this to false, then the resource groups will not be deleted.  This is
+    expected behavior when combined with the `azd down` command which will take responsibility for deleting
+    the resource groups.
 .NOTES
     This command requires that Az modules are installed and imported. It also requires that you have an
     active Azure session.  If you are not authenticated with Azure, you will be prompted to authenticate.
@@ -32,7 +36,10 @@ Param(
     [Parameter(Mandatory = $false)][string]$SecondaryResourceGroup,
     [Parameter(Mandatory = $false)][string]$SpokeResourceGroup,
     [Parameter(Mandatory = $false)][string]$SecondarySpokeResourceGroup,
-    [Parameter(Mandatory = $false)][string]$HubResourceGroup
+    [Parameter(Mandatory = $false)][string]$HubResourceGroup,
+    [Parameter(Mandatory = $false)][switch]$SkipResourceGroupDeletion,
+    [Parameter(Mandatory = $false)][switch]$Purge,
+    [Parameter(Mandatory = $false)][switch]$NoPrompt
 )
 
 
@@ -77,7 +84,11 @@ $rgSpoke = ""
 $rgHub = ""
 $rgSecondaryApplication = ""
 $rgSecondarySpoke = ""
+$rgContainerAppEnvironment = ""
+$rgSecondaryContainerAppEnvironment = ""
 #$CleanupAzureDirectory = $false
+
+$azdConfig = azd env get-values -o json | ConvertFrom-Json -Depth 9 -AsHashtable
 
 if ($Prefix) {
     $rgPrefix = $Prefix
@@ -92,7 +103,6 @@ if ($Prefix) {
             "No .azure directory found and no resource group information provided - cannot clean up"
             exit 8
         }
-        $azdConfig = azd env get-values -o json | ConvertFrom-Json -Depth 9 -AsHashtable
         $environmentName = $azdConfig['AZURE_ENV_NAME']
         $environmentType = $azdConfig['AZURE_ENV_TYPE'] ?? 'dev'
         $location = $azdConfig['AZURE_LOCATION']
@@ -102,7 +112,7 @@ if ($Prefix) {
         $rgSpoke = "$rgPrefix-$location-spoke"
         $rgSecondaryApplication = "$rgPrefix-$locationSecondary-2-application"
         Write-Host "Secondary Application Resource Group: $rgSecondaryApplication"
-        $rgSecondarySpoke = "$rgPrefix-$locationSecondary-2-spoke"    
+        $rgSecondarySpoke = "$rgPrefix-$locationSecondary-2-spoke"
         Write-Host "Secondary Spoke Resource Group: $rgSecondarySpoke"
         $rgHub = "$rgPrefix-hub"
         #$CleanupAzureDirectory = $true
@@ -164,9 +174,7 @@ function Get-AzBudget($resourceGroupName) {
     return $result.value
 }
 
-# Removed all budgets that are scoped to a resource group of interest.
 function Remove-ConsumptionBudgetForResourceGroup($resourceGroupName) {
-    # Get-AzConsumptionBudget -ResourceGroupName $resourceGroupName
     Get-AzBudget -ResourceGroupName $resourceGroupName
     | Foreach-Object {
         "`tRemoving $resourceGroupName::$($_.name)" | Write-Output
@@ -201,6 +209,62 @@ function Remove-ResourceGroupFromAzure($resourceGroupName) {
     }
 }
 
+function Test-EntraAppRegistrationExists($name) {
+    $appRegistration = Get-AzADApplication -DisplayName $name -ErrorAction SilentlyContinue
+    return $null -ne $appRegistration
+}
+
+function Remove-AzADApplicationByName($name) {
+    $appRegistration = Get-AzADApplication -DisplayName $name -ErrorAction SilentlyContinue
+    if ($appRegistration) {
+        "`tRemoving $name" | Write-Output
+        Remove-AzADApplication -ObjectId $appRegistration.Id
+    }
+}
+
+function Get-ResourceToken($resourceGroupName) {
+    $defaultRedisNamePrefix = 'redis-'
+    $redisInstances = Get-AzRedisCache -ResourceGroupName $resourceGroupName -ErrorAction SilentlyContinue
+
+    if ($redisInstances.Count -eq 0) {
+        return "notfound"
+    }
+
+    return ($redisInstances | Select-Object -First 1).Name.Substring($defaultRedisNamePrefix.Length)
+}
+
+<#
+.SYNOPSIS
+    Reads input from the user, but taking care of default value and request to
+    not prompt the user.
+.PARAMETER Prompt
+    The prompt to display to the user.
+.PARAMETER DefaultValue
+    The default value to use if the user just hits Enter.
+.PARAMETER NoPrompt
+    If specified, don't prompt - just use the default value.
+#>
+function Read-ApplicationPrompt {
+    param(
+        [Parameter(Mandatory = $true)]
+        [string] $Prompt,
+
+        [Parameter(Mandatory = $true)]
+        [string] $DefaultValue,
+
+        [Parameter(Mandatory = $false)]
+        [switch] $NoPrompt = $false
+    )
+
+    $returnValue = ""
+    if (-not $NoPrompt) {
+        $returnValue = Read-Host -Prompt "`n$($Prompt) [default: $(Get-HighlightedText($DefaultValue))] "
+    }
+    if ([string]::IsNullOrWhiteSpace($returnValue)) {
+        $returnValue = $DefaultValue
+    }
+    return $returnValue
+}
 "`nCleaning up environment for application '$rgApplication'" | Write-Output
 
 # Get the list of resource groups to deal with
@@ -208,6 +272,13 @@ $resourceGroups = [System.Collections.ArrayList]@()
 if (Test-ResourceGroupExists -ResourceGroupName $rgApplication) {
     "`tFound application resource group: $rgApplication" | Write-Output
     $resourceGroups.Add($rgApplication) | Out-Null
+    
+    $resourceToken=(Get-ResourceToken -resourceGroupName $rgApplication) # expecting to be something like 'fjmjdbizcdxt4'
+    $possibleContainerAppEnvironmentGroup = "ME_acae-common-$resourceToken"
+    if (Test-ResourceGroupExists -ResourceGroupName $possibleContainerAppEnvironmentGroup) {
+        "`tFound container app environment resource group: $possibleContainerAppEnvironmentGroup" | Write-Output
+        $resourceGroups.Add($possibleContainerAppEnvironmentGroup) | Out-Null
+    }
 } else {
     "`tConfirm the correct subscription was selected and check the spelling of the group to be deleted" | Write-Warning
     "`tCould not find resource group: $rgApplication" | Write-Error
@@ -216,6 +287,13 @@ if (Test-ResourceGroupExists -ResourceGroupName $rgApplication) {
 if (Test-ResourceGroupExists -ResourceGroupName $rgSecondaryApplication) {
     "`tFound secondary application resource group: $rgSecondaryApplication" | Write-Output
     $resourceGroups.Add($rgSecondaryApplication) | Out-Null
+
+    $resourceToken=(Get-ResourceToken -resourceGroupName $rgSecondaryApplication) # expecting to be something like 'fjmjdbizcdxt4'
+    $possibleContainerAppEnvironmentGroup = "ME_acae-common-$resourceToken"
+    if (Test-ResourceGroupExists -ResourceGroupName $possibleContainerAppEnvironmentGroup) {
+        "`tFound secondary container app environment resource group: $possibleContainerAppEnvironmentGroup" | Write-Output
+        $resourceGroups.Add($possibleContainerAppEnvironmentGroup) | Out-Null
+    }
 }
 
 
@@ -232,9 +310,57 @@ if (Test-ResourceGroupExists -ResourceGroupName $rgHub) {
     $resourceGroups.Add($rgHub) | Out-Null
 }
 
+$resourceToken=(Get-ResourceToken -resourceGroupName $rgApplication) # expecting to be something like 'fjmjdbizcdxt4'
+$appRegistrations = [System.Collections.ArrayList]@()
+$calculatedAppRegistrationNameForApi = "$rgPrefix-api-webapp-$resourceToken".Substring(3)
+$calculatedAppRegistrationNameForFrontend = "$rgPrefix-front-webapp-$resourceToken".Substring(3)
+
+if (Test-EntraAppRegistrationExists -Name $calculatedAppRegistrationNameForApi) {
+    "`tFound Entra ID App Registration: $calculatedAppRegistrationNameForApi" | Write-Output
+    $appRegistrations.Add($calculatedAppRegistrationNameForApi) | Out-Null
+}
+if (Test-EntraAppRegistrationExists -Name $calculatedAppRegistrationNameForFrontend) {
+    "`tFound Entra ID App Registration: $calculatedAppRegistrationNameForFrontend" | Write-Output
+    $appRegistrations.Add($calculatedAppRegistrationNameForFrontend) | Out-Null
+}
+
+# Determine if we need to purge the App Configuration and Key Vault.
+$defaultPurgeResources = if ($Purge) { "y" } else { "n" }
+$purgeResources = Read-ApplicationPrompt -Prompt "Do you wish to puge resources that cannot be reassigned immediately (such as Key Vault)? [y/n]" -DefaultValue $defaultPurgeResources -NoPrompt:$NoPrompt
+
 # press enter to proceed
-"`nPress enter to proceed with cleanup or CTRL+C to cancel" | Write-Output
-$null = Read-Host
+if (-not $NoPrompt) {
+    "`nPress enter to proceed with cleanup or CTRL+C to cancel" | Write-Output
+    $null = Read-Host
+}
+
+# we don't want to delete the app registrations because we reuse them when running in pipeline
+# when running in pipeline, the AZURE_PRINCIPAL_TYPE is set to 'ServicePrincipal'
+if ($azdConfig['AZURE_PRINCIPAL_TYPE'] -eq 'User') {    
+    "`nRemoving Entra ID App Registration..." | Write-Output
+    foreach($appRegistration in $appRegistrations) {
+        Remove-AzADApplicationByName -Name $appRegistration
+    }
+}
+
+if ($purgeResources -eq "y") {
+    "> Remove and purge purgeable resources:" | Write-Output
+    foreach ($resourceGroupName in $resourceGroups) {
+        Get-AzKeyVault -ResourceGroupName $resourceGroupName | Foreach-Object {
+            "`tRemoving $($_.VaultName)" | Write-Output
+            Remove-AzKeyVault -VaultName $_.VaultName -ResourceGroupName $resourceGroupName -Force
+            "`tPurging $($_.VaultName)" | Write-Output
+            Remove-AzKeyVault -VaultName $_.VaultName -Location $_.Location -InRemovedState -Force -ErrorAction SilentlyContinue
+        }
+
+        Get-AzAppConfigurationStore -ResourceGroupName $resourceGroupName | Foreach-Object {
+            "`tRemoving $($_.Name)" | Write-Output
+            Remove-AzAppConfigurationStore -Name $_.Name -ResourceGroupName $resourceGroupName
+            "`tPurging $($_.Name)" | Write-Output
+            Clear-AzAppConfigurationDeletedStore -Location $_.Location -Name $_.Name -ErrorAction SilentlyContinue
+        }
+    }
+}
 
 "`nRemoving resources from resource groups..." | Write-Output
 "> Private Endpoints:" | Write-Output
@@ -252,16 +378,15 @@ foreach ($resourceGroupName in $resourceGroups) {
     Remove-DiagnosticSettingsForResourceGroup -ResourceGroupName $resourceGroupName
 }
 
-"`nRemoving resource groups in order..." | Write-Output
-Remove-ResourceGroupFromAzure -ResourceGroupName $rgApplication
-Remove-ResourceGroupFromAzure -ResourceGroupName $rgSecondaryApplication
-Remove-ResourceGroupFromAzure -ResourceGroupName $rgSpoke
-Remove-ResourceGroupFromAzure -ResourceGroupName $rgSecondarySpoke
-Remove-ResourceGroupFromAzure -ResourceGroupName $rgHub
-
-# if ($CleanupAzureDirectory -eq $true -and (Test-Path -Path ./.azure -PathType Container)) {
-#     "Cleaning up Azure Developer CLI state files." | Write-Output
-#     Remove-Item -Path ./.azure -Recurse -Force
-# }
-
-"`nCleanup complete." | Write-Output
+# if $SkipResourceGroupDeletion is false, then we skip the resource group deletion
+# flag is expected to be set to false when combined with the `azd down` command
+if (-not $SkipResourceGroupDeletion) {
+    "`nRemoving resource groups in order..." | Write-Output
+    Remove-ResourceGroupFromAzure -ResourceGroupName $rgApplication
+    Remove-ResourceGroupFromAzure -ResourceGroupName $rgSecondaryApplication
+    Remove-ResourceGroupFromAzure -ResourceGroupName $rgSpoke
+    Remove-ResourceGroupFromAzure -ResourceGroupName $rgSecondarySpoke
+    Remove-ResourceGroupFromAzure -ResourceGroupName $rgHub
+    
+    "`nCleanup complete." | Write-Output
+}
